Improvements to clients found from attempting a hyperdriver integration.
